templates/MicroserviceDddCqrs/Src/Core/Application/Products/Commands/CreateProduct/CreateProductCommandHandler.cs
using System.Threading;
using System.Threading.Tasks;
using Application.Common.Exceptions;
using Application.Common.Interfaces;
using Application.Products.Models;
using AutoMapper;
using Domain.Entities;
using MediatR;
using Microsoft.EntityFrameworkCore;

namespace Application.Products.Commands.CreateProduct
{
    public class CreateProductCommandHandler : IRequestHandler<CreateProductCommand, int>
    {
        private readonly IAppDbContext _context;
        private readonly IMapper _mapper;

        public CreateProductCommandHandler(IAppDbContext context, IMapper mapper)
        {
            _context = context;
            _mapper = mapper;
        }

        public async Task<int> Handle(CreateProductCommand request, CancellationToken cancellationToken)
        {
            var anyProductWithSameName = await _context.Products.AnyAsync(p => p.Name == request.Name, cancellationToken);
            if (anyProductWithSameName)
                throw new ProductNameAlreadyInUseException($"Product with name: {request.Name} already exist");

            var product = _mapper.Map<Product>(request);

            await _context.Products.AddAsync(product, cancellationToken);

            await _context.SaveChangesAsync(cancellationToken);

            return product.Id;
        }
    }
}